requirements.txt
#ml refereces
pandas==1.2.4
tensorflow-gpu==2.5.0
tokenizers==0.10.3
transformers==4.9.2
numpy==1.19.5
scikit-learn==0.24.1
spacy==3.0.6
click==7.1.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0.tar.gz#egg=en_core_web_md
pylint==2.8.2
lime==0.2.0.1
anchor-exp==0.0.2.0
# web references
python-dotenv==0.17.1
django==3.2.4
django-create-react-app==0.8.4
channels==3.0.3
django-proxy==1.2.1
requests==2.25.1
azure-storage-blob==12.8.1
json-lines==0.5.0
# live webserver
gunicorn==20.1.0
uvicorn==0.14.0
websockets==9.1